+++ packages/core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0
 
 - New block API ([#125](https://github.com/andywer/webpack-blocks/issues/125))
+- Added `match()`
 - Added `text/html` file type
 - Dropped Babel, requires node 6+