+++ packages/webpack/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @webpack-blocks/webpack - Changelog
 
+## Next release
+
+- Added `match()`
+- Added `resolve()`, deprecate `resolveAliases()`
+
 ## 1.0
 
 - Updated for new core API ([#125](https://github.com/andywer/webpack-blocks/issues/125))